+++ otj-metrics-core/src/main/java/com/opentable/metrics/AtomicDoubleGauge.java
