+++ otj-metrics-core/src/main/java/com/opentable/metrics/AtomicLongGauge.java
