+++ otj-metrics-core/src/main/java/com/opentable/metrics/DefaultMetricsConfiguration.java
