+++ otj-metrics-core/src/main/java/com/opentable/metrics/FloatingPointHistogram.java
