+++ otj-metrics-core/src/main/java/com/opentable/metrics/health/MediocreHealthCheck.java
