+++ otj-metrics-core/src/main/java/com/opentable/metrics/http/CheckState.java
