+++ otj-metrics-core/src/main/java/com/opentable/metrics/http/HealthCheckResponse.java
