+++ otj-metrics-core/src/main/java/com/opentable/metrics/jvm/CpuLoadGauge.java
