+++ otj-metrics-jaxrs/src/main/java/com/opentable/metrics/http/HealthHttpConfiguration.java
