+++ otj-metrics-jaxrs/src/test/java/com/opentable/metrics/http/HealthApiTest.java
