+++ otj-metrics/pom.xml
@@ -19,30 +19,21 @@
 
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.opentable</groupId>
-        <artifactId>otj-parent-spring</artifactId>
-        <version>173</version>
+        <groupId>com.opentable.components</groupId>
+        <artifactId>otj-metrics-parent</artifactId>
+        <version>2.6.13-SNAPSHOT</version>
     </parent>
-
-    <scm>
-        <connection>scm:git:git://github.com/opentable/otj-metrics.git</connection>
-        <developerConnection>scm:git:git@github.com:opentable/otj-metrics.git</developerConnection>
-        <url>http://github.com/opentable/otj-metrics</url>
-        <tag>HEAD</tag>
-    </scm>
-
-    <groupId>com.opentable.components</groupId>
-    <artifactId>otj-metrics-parent</artifactId>
-    <name>OpenTable Metrics - Parent</name>
-    <version>2.6.13-SNAPSHOT</version>
+    <artifactId>otj-metrics</artifactId>
+    <name>OpenTable Metrics - Legacy Relocation Artifact</name>
+    <description>Legacy - OpenTable Metrics with JAX-RS endpoints</description>
     <packaging>pom</packaging>
-    <description>OpenTable Metrics - Parent</description>
 
-    <modules>
-        <module>otj-metrics-core</module>
-        <module>otj-metrics-jaxrs</module>
-        <module>otj-metrics-mvc</module>
-        <module>otj-metrics</module>
-    </modules>
+    <distributionManagement>
+        <relocation>
+            <artifactId>otj-metrics-jaxrs</artifactId>
+            <!-- This message doesn't appear anywhere, oddly -->
+            <message>Please use otj-metrics-core and otj-metrics-jaxrs instead!</message>
+        </relocation>
+    </distributionManagement>
 
 </project>
